# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arius API",
    "version": "0.1.0",
    "description": "Documentation OpenAPI de l'API Arius CRM: authentification, entreprises, contacts, notes, RDVs, devis, objectifs, chiffre d'affaires, profil et export RGPD."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Message d'erreur"
          }
        }
      },
...
      "EntrepriseInput": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "example": "Acme SAS"
          },
          "secteur": {
            "type": "string",
            "example": "Industrie"
          },
          "adresse": {
            "type": "string",
            "example": "10 rue de Paris"
          },
          "telephone": {
            "type": "string",
            "example": "+33123456789"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "site_web": {
            "type": "string",
            "example": "https://example.com"
          }
        }
      },
...
      "ObjectifInput": {
        "type": "object",
        "properties": {
          "mois": {
            "type": "string",
            "example": "2026-06"
          },
          "montant": {
            "type": "number",
            "example": 50000
          }
        }
      },
      "CAInput": {
        "type": "object",
        "properties": {
          "entreprise_id": {
            "type": "string"
          },
          "montant": {
            "type": "number",
            "example": 12000
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentification requise ou token invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/v1/upload": {
      "post": {
        "summary": "Upload le logo d'une entreprise",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "entrepriseId",
                  "image"
                ],
                "properties": {
                  "entrepriseId": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logo uploade"
          },
          "400": {
            "description": "Payload invalide"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/objectifs": {
      "get": {
        "summary": "Liste les objectifs",
        "tags": [
          "Objectifs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des objectifs"
          }
        }
      },
      "post": {
        "summary": "Cree un objectif",
        "tags": [
          "Objectifs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectifInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Objectif cree"
          }
        }
      }
    },
...
    "/v1/entreprises": {
      "get": {
        "summary": "Liste les entreprises de l'utilisateur",
        "tags": [
          "Entreprises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des entreprises"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Cree une entreprise",
        "tags": [
          "Entreprises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntrepriseInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entreprise creee"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/ca": {
      "get": {
        "summary": "Liste les entrees de chiffre d'affaires",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste du chiffre d'affaires"
          }
        }
      },
      "post": {
        "summary": "Cree une entree de chiffre d'affaires",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CAInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entree creee"
          }
        }
      }
    },
    "/v1/ca/stats": {
      "get": {
        "summary": "Recupere les statistiques de chiffre d'affaires",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques CA"
          }
        }
      }
    },
    "/v1/ca/entreprise/{entreprise_id}": {
      "get": {
        "summary": "Recupere le chiffre d'affaires d'une entreprise",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entreprise_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CA de l'entreprise"
          }
        }
      }
    },
    "/v1/ca/{id}": {
      "put": {
        "summary": "Met a jour une entree de chiffre d'affaires",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CAInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entree mise a jour"
          }
        }
      },
      "delete": {
        "summary": "Supprime une entree de chiffre d'affaires",
        "tags": [
          "CA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entree supprimee"
          }
        }
      }
    },
...
  }
}